v1.2.1 rev-6 | 2024/07/23 15:19 | Added Dekker Floats and refactored Float80/Float128 handeling.
Added double Dekker float implementations for Float32, Float64, and Float80. Float64x2 is roughly 4.3x faster  than Float128 with similar precision. In the future, I hope to add SSE2/AVX versions of the Float32/Float64 double Dekker floats. Float80.hpp and Float128.hpp were added to make it easier to include __float80/long double and __float128/quadmath. A Float64 OpenCL/GPU rendering option has been added; I am yet to add the Dekker floats to OpenCL rendering though.

+++ src/frame_Transformation.cpp
@@ -20,7 +20,7 @@
 #include <xmmintrin.h>
 
 // Precision used for transformations
-typedef fp32 fpTran;
+typedef fp64 fpTran;
 
 /**
  * @brief Generic Method. Runs faster, but leaves pixels gaps and skips pixels near the edge of the frame.
